Optimized client rendering
Now, instead of redrawing the whole stroke stack on each stroke, the
user's temporary stroke is drawn to the viewport while the buffer gets
the final smoothed stroke.
Also, the canvas is compressed to a DataURL and stored in local storage,
and is read from on subsequent page loads for a much, much faster user
experience. Right now, the stack is still being stored locally.